% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan_plot.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plot}
\usage{
manhattan_plot(
  x,
  chrom.colors,
  chrLabels,
  genomewideline = 5,
  highlight = NULL,
  is.logp = TRUE,
  ...
)
}
\arguments{
\item{x}{Dataframe of GWAS results. Requires columns named: "snp", "chr", "bp", "p".}

\item{chrom.colors}{Chromosome colors to alternate}

\item{highlight}{Logical. Highlight character vector of snps in x}

\item{threshold}{Genome-wide significance threshold}

\item{p}{p-value}
}
\value{
A Manhattan plot of GWAS results
}
\description{
This function takes a dataframe of GWAS results as input, which
requires columns named: "snp", "chr", "bp", and "p", and returns
a Manhattan plot. This function assumes snp position is in
base pairs "bp". Options are: genome-wide significance threshold,
chromosome colors, and snps to highlight.
}
