% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ-plot}
\usage{
qq_plot(pvector, traitName = "", ...)
}
\arguments{
\item{pvector}{Vector of p-values}

\item{traitName}{Name of trait to plot.}
}
\value{
A QQ plot of GWAS results
}
\description{
This function takes a vector of pvalues from GWAS results as input,
and returns a QQ-plot
}
